chore(flashblocks): modularize types
